+++ modules/renter/uploadheap.go
@@ -426,8 +426,9 @@ func (r *Renter) managedBuildUnfinishedChunk(entry *filesystem.FileNode, chunkIn
 		physicalChunkData: make([][]byte, entry.ErasureCode().NumPieces()),
 		verifyPieceRoot:   make([]crypto.Hash, entry.ErasureCode().NumPieces()),
 
-		pieceUsage:  make([]bool, entry.ErasureCode().NumPieces()),
-		unusedHosts: make(map[string]struct{}, len(hosts)),
+		availableChan: make(chan struct{}),
+		pieceUsage:    make([]bool, entry.ErasureCode().NumPieces()),
+		unusedHosts:   make(map[string]struct{}, len(hosts)),
 	}
 
 	// Every chunk can have a different set of unused hosts.