+++ CHANGELOG.md
@@ -9,6 +9,19 @@ Sia uses the following versioning scheme, vX.X.X.X
 Version History
 ---------------
 
+Latest:
+
+### v1.4.2.1
+
+**Key Updates**
+ - Wallet can generate an address before it finishes scanning the blockchain
+ - FUSE folders can now be mounted with 'AllowOther' as an option
+
+**Bugs Fixed**
+ - Unable to upload a new file if 'force' is set and no file exists to delete.
+ - Siac would not always delete a file or folder correctly
+ - Divide by zero error when setting the allowance with an empty period
+
 Dec 2019:
 
 ### v1.4.2.0
@@ -35,7 +48,7 @@ Dec 2019:
  - Gateway bandwidth monitoring
  - Ability to pause uploads/repairs
 
-**Bugs**
+**Bugs Fixed**
  - Missing return statements in API (http: superfluous response.WriteHeader call)
  - Stuck Loop fixes (chunks not being added due to directory siapath never being set)
  - Rapid Cycle repair loop on start up